Switch to lexicographic sorting of releases
LooseVersion causes the sort-order of 'releases' list to put the nightly at the end of the list and the most recent release as second-from-end-of-list.  The rest of the code expects nightly to be at the beginning of the list (releases[0]) and the latest release to be at the end of the list (releases[-1]).

This change eliminates LooseVersion as the sort key in favor of default lexicographic sort, resulting in the expected positions of nightly and latest release.